+++ Sources/PitchSpeller/ModifierDirection.swift
@@ -1,0 +1,14 @@
+//
+//  ModifierDirection.swift
+//  PitchSpeller
+//
+//  Created by James Bean on 6/10/18.
+//
+
+/// The possible directions of `QuarterStepModifier` values within the context of
+/// a `PitchSpellingCategoryProtocol`.
+enum ModifierDirection: Hashable {
+    case down
+    case neutral
+    case up
+}
